chore: add redis v2 health check
