+++ server/src/init.ts
@@ -31,6 +31,10 @@ import {
 	stopRedisMonitor,
 	warmupRegionalRedis,
 } from "./external/redis/initRedis.js";
+import {
+	startRedisV2Monitor,
+	stopRedisV2Monitor,
+} from "./external/redis/initUtils/redisV2Availability.js";
 import { createHonoApp } from "./initHono.js";
 import { otelSdk } from "./instrumentation.js";
 import { checkEnvVars } from "./utils/initUtils.js";
@@ -50,6 +54,7 @@ const init = async ({ startupStartedAt }: { startupStartedAt: number }) => {
 	initPgHealthMonitor({ client: clientCritical });
 
 	startRedisMonitor();
+	startRedisV2Monitor();
 	void warmupRegionalRedis().catch((error) => {
 		logger.warn("[Redis] Warmup failed", { error });
 	});
@@ -150,6 +155,7 @@ async function gracefulShutdown() {
 		}
 		shutdownPgHealthMonitor();
 		stopRedisMonitor();
+		stopRedisV2Monitor();
 		stopAllEdgeConfigPolling();
 		await Promise.all([
 			client.end(),
